Make controller independent from openstack
