+++ internal/openstack/port_manager.go
@@ -62,14 +62,6 @@ func (opts CustomCreateOpts) ToPortCreateMap() (map[string]interface{}, error) {
 	return gophercloud.BuildRequestBody(opts, "port")
 }
 
-type L3PortManager interface {
-	ProvisionPort() (string, error)
-	CleanUnusedPorts(usedPorts []string) error
-	GetAvailablePorts() ([]string, error)
-	GetExternalAddress(portID string) (string, string, error)
-	GetInternalAddress(portID string) (string, error)
-}
-
 type OpenStackL3PortManager struct {
 	client    *gophercloud.ServiceClient
 	networkID string
